use viewmodel to get user id

+++ app/src/main/java/com/example/vetapp/Database/Entities/Reports.kt
@@ -15,6 +15,6 @@ import androidx.room.PrimaryKey
 data class Reports (
     @PrimaryKey (autoGenerate = true) val Id : Int = 0,
     var name : String,
-    var userId : Int
+    var userId : Long
 )
